Decouple the code that executes checks from the agent
