+++ agent/checks/docker_windows.go
@@ -1,3 +1,3 @@
-package agent
+package checks
 
 const DefaultDockerHost = "npipe:////./pipe/docker_engine"
