+++ agent/exec/exec_windows.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package agent
+package exec
 
 import (
 	"os"
@@ -9,8 +9,8 @@ import (
 	"syscall"
 )
 
-// ExecScript returns a command to execute a script through a shell.
-func ExecScript(script string) (*exec.Cmd, error) {
+// Script returns a command to execute a script through a shell.
+func Script(script string) (*exec.Cmd, error) {
 	shell := "cmd"
 	if other := os.Getenv("SHELL"); other != "" {
 		shell = other
