+++ agent/util.go
@@ -92,15 +92,6 @@ GROUP:
 	return nil
 }
 
-// ExecSubprocess returns a command to execute a subprocess directly.
-func ExecSubprocess(args []string) (*exec.Cmd, error) {
-	if len(args) == 0 {
-		return nil, fmt.Errorf("need an executable to run")
-	}
-
-	return exec.Command(args[0], args[1:]...), nil
-}
-
 // ForwardSignals will fire up a goroutine to forward signals to the given
 // subprocess until the shutdown channel is closed.
 func ForwardSignals(cmd *exec.Cmd, logFn func(error), shutdownCh <-chan struct{}) {
